:sparkles: Show related event logs on resource page
